fix!: Update user/group IDs to 1568 and add HEALTHCHECK NONE
Changed nonroot user and group IDs from 568 to 1568 to resolve potential conflicts with existing IDs. Updated the Dockerfile to include `HEALTHCHECK NONE` to disable health checks. Adjusted the README to reflect the new user and group IDs.

BREAKING CHANGE: Changing the UID because 1000 is where the user IDs start, and we may accidentally conflict with something on the host